Safari-609.1.11.1/LayoutTests/css3/font-variant-parsing-expected.txt

PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-ligatures').cssText is "normal"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-ligatures').cssText is "normal"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-ligatures').cssText is "normal"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-ligatures').cssText is "normal"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-ligatures').cssText is "common-ligatures"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-ligatures').cssText is "normal"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-ligatures').cssText is "no-common-ligatures"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-ligatures').cssText is "discretionary-ligatures"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-ligatures').cssText is "no-discretionary-ligatures"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-ligatures').cssText is "historical-ligatures"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-ligatures').cssText is "no-historical-ligatures"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-ligatures').cssText is "contextual"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-ligatures').cssText is "no-contextual"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-ligatures').cssText is "common-ligatures no-contextual"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-ligatures').cssText is "common-ligatures no-contextual"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-ligatures').cssText is "normal"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-ligatures').cssText is "none"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-ligatures').cssText is "none"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-position').cssText is "normal"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-position').cssText is "normal"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-position').cssText is "normal"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-position').cssText is "sub"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-position').cssText is "super"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-position').cssText is "normal"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-caps').cssText is "normal"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-caps').cssText is "normal"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-caps').cssText is "normal"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-caps').cssText is "normal"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-caps').cssText is "normal"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-caps').cssText is "small-caps"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-caps').cssText is "all-small-caps"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-caps').cssText is "petite-caps"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-caps').cssText is "all-petite-caps"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-caps').cssText is "unicase"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-caps').cssText is "titling-caps"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-caps').cssText is "normal"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-caps').cssText is "normal"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-numeric').cssText is "normal"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-numeric').cssText is "normal"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-numeric').cssText is "normal"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-numeric').cssText is "normal"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-numeric').cssText is "normal"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-numeric').cssText is "lining-nums"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-numeric').cssText is "oldstyle-nums"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-numeric').cssText is "proportional-nums"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-numeric').cssText is "tabular-nums"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-numeric').cssText is "diagonal-fractions"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-numeric').cssText is "stacked-fractions"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-numeric').cssText is "ordinal"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-numeric').cssText is "slashed-zero"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-numeric').cssText is "lining-nums slashed-zero"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-numeric').cssText is "normal"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-numeric').cssText is "normal"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-numeric').cssText is "normal"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-numeric').cssText is "lining-nums proportional-nums"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-numeric').cssText is "lining-nums diagonal-fractions"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-numeric').cssText is "ordinal slashed-zero"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-alternates').cssText is "normal"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-alternates').cssText is "normal"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-alternates').cssText is "normal"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-alternates').cssText is "normal"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-alternates').cssText is "historical-forms"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-alternates').cssText is "normal"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-east-asian').cssText is "normal"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-east-asian').cssText is "normal"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-east-asian').cssText is "normal"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-east-asian').cssText is "normal"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-east-asian').cssText is "jis78"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-east-asian').cssText is "jis83"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-east-asian').cssText is "jis90"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-east-asian').cssText is "jis04"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-east-asian').cssText is "simplified"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-east-asian').cssText is "traditional"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-east-asian').cssText is "full-width"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-east-asian').cssText is "proportional-width"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-east-asian').cssText is "ruby"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-east-asian').cssText is "jis83"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-east-asian').cssText is "traditional"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-east-asian').cssText is "traditional full-width"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-east-asian').cssText is "jis04 proportional-width"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-east-asian').cssText is "jis04 proportional-width ruby"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-east-asian').cssText is "jis83 ruby"
PASS window.getComputedStyle(document.getElementById('testElement')).getPropertyCSSValue('font-variant-east-asian').cssText is "proportional-width"
PASS successfullyParsed is true

TEST COMPLETE
